+++ src/main/java/hr/btb/testapi/rowMapper/AdresaRowMapper.java
@@ -12,7 +12,7 @@ public class AdresaRowMapper implements RowMapper<Adresa> {
 
 	public Adresa mapRow(ResultSet rs, int rowNum) throws SQLException {
 			Adresa adresa = new Adresa();
-			adresa.setId(rs.getLong("id"));
+			adresa.setId(rs.getInt("id"));
 			adresa.setUlica(rs.getString("ulica"));
 			adresa.setKucni_broj(rs.getString("kucni_broj"));
 			adresa.setGrad(rs.getString("grad"));
